URL: In Tree.hpp we cannot store objects of Node - must use pointers.
Or rather, we can't store objects and then keep references to them
because underneath, vector<> may reallocate and reshuffle it's RAM.

Sadly, we must use pointers then and allocate on the heap as usual.

We could store the tree IDs of the children but then Node we need to
know about Tree.