utility/Node: Allow Node to be used as a base class (as intended).
Node should be able to be extended - classes that derived from it
should override new_node() to be sure to create the correct type.

The crawler scheduler will override Node to attach various scheduling
properties/policies to each leaf node and likely store a pointer to
the parsed URL object.