url/URL: Provide an interface to Page, Path and Domain for id().
Internally we may use a hashing method for creating a unique
URL fragment/component ID for the page, path and domain.

This will allow us to use said attributes in hash tables etc.
for fast retrieval.

IDContainer provides this abstraction layer. Internally, content
must point to (const ref) the actual attribute value in Attribute.

When hashing a URL we break it down into the individual hash
of a page, path and domain so we can identify each fragment
on it's own. However, when hashing say path we will need to provide
the domain hash as the seed (to extend the hash as it were) and
similarly when hashing the page we wish to seed it with the path.