utility/Node: Maintain find() interface but separate implementation.
The existing binary search (D&Q) implementation is now in bsearch()
allowing the underlying implementation to be changed if necessary.

Also, with the advent of OpenMP we can split a binary search across
processors. Compile with -DHAVE_OPENMP.

This is still largely expermiental as it is not clear that parallelising
here gives any benefit in scalability, performance etc.