url: Distinguish between IP octects and domain components during parsing.
TreeIterator must iterate forwards through IP addresses but backward
through domain names in the tree - so we must know what lies in the
domain container. Do this as we parse so we don't add further complexity
in making a final pass through the domain container to scan it for a set
of integers rather than strings.