utility/Node: Remove dependency on templated Node<T> for basic operations
I don't want to be dealing with types when I don't need them! In order
to provide more flexibility break away from the templated Node<> as
the starting base class and employ a more basic, generic base.

Similarly we do the same for Tree. This will allow us to support
basic processing of trees and nodes without having to know their
data type (for the key/value) when we don't need it.