utility/NodeIO: Provide print functors in the namespace io.
So we can print out different representations of the tree
in memory make use of some print functors - we don't want
to bloat the tree/node structures with different print code.

Here, as well as the usual ASCII art tree, we now present
output in the form of the DOT language. Using DOT we can
get a graphical layout of our tree structure.