sched/Scheduler: Make use of the BreadCrumbTrail class.
At each selection of a node in traverse_branch() we store the X,Y point
of the node. Later, with replay_run(), we can play back the last schedule run
onto a stream - in DOT format.

Eventually this will be moved out of the core scheduler work. It is here
for simplicity to beginwith but doesn't really belong here. It should be
in NodeIO like the other print functors.