utility:/ChunkManager: Manage n Chunk objects.
ChunkManager, by pooling as many chunks as necessary, is not restricted
by the fixed size of a Chunk. It holds as many Chunks as there is RAM
available, if necessary. It grows it's internal store (the no. of chunks)
by 2 each time a new allocation is needed. This is to try and reduce the
no. of calls to the underlying Chunk malloc.