sched/Scheduler: Use a hash table to keep a fast store of page ID to Node.
When the Scheduler receives confirmation from a crawler that it has
successfully crawled (etc.) a page we must update the weight etc. of
the node that holds that page. Therefore we compute the hash of a page
and store that as the key to the node for a quick, O(1), retrieval of
the node. We don't wish to walk the tree for that - only when choosing
our next schedule.

We use tr1::unordered_map as that uses tr1::hashtable beneath it's simple
interface. Because we already compute a hash of various segments of a URL
be sure to use those - we cannot store a map (std or tr1) of K,V where K is,
 say either Page or URL because when we send details of a crawl to the crawler
 we send the bare minimum. I.e. the integer hash of the object and a string
 of the URL.