utility/subs: Minimal, simple observer pattern.
Using three simple classes; Event, Publisher and Subscriber we now have
an elementry subscribe/publish system based on the observer pattern.

Derive from Event to build your own event object and register it with a
publisher. Derive from Subscriber for objects interested when an event
changes and add them to the publisher. Call broadcast() when subscribers
should be made aware of the event change.