sched/Scheduler: Add bare-bones Subscriber support.
We want the scheduler to be informed when Crawlers have returned
results about the pages that were scheduled to them so the scheduler
can quickly update the tree (weigh branches) again.

We can remove the comment about boost.signals at it seemed unnecessary
and overkill here. Something simpler is more appropriate at this stage.