net/Endpoint: Generic (though TCP) endpoint for client/server use.
Implements the raw socket-level inbound and outbound data transfer.
Makes exclusive use of the Boost.ASIO TCP sockets and the associated
asynchronous read/write calls handled by the io_service and the
provided callbacks below.

Use Endpoint::create() to return a new reference counted Connection object.

Endpoint relies on a ProtocolHandler for understanding the wire
protocol - it just deals with raw bytes beneath but depends on the
interface ProtocolHandler presents (and expects the implementation
to deal with the specific protocol).