utility/events: Optionally pass a pointer to a boost::io_service.
By doing so broadcast() becomes non-blocking as all receive()
invocations are in fact dispatched to the io_service's queue.

Internally, io_service has been typedef'd to Proactor - instantiate
it this way if you wish. Pass in an io_service that you may already
be using (say, in a pool of threads for running run()) and the
events system suddenly becomes asynchronous - broadcast() will
return immediately and the receive methods will be executed from
within the io_service threads.

broadcast() now returns the number of non-blocking calls it
dispatched or 0 if it blocks.