test/protocol-handler: The call buffer2message should return used bytes.
Previously this call returned the remainder of the buffer that it didn't use.
In order provide a consistent interface this now returns the number of used
bytes as does message2buffer.

Do not send 0-byte files. For our test protocol this is unnecessary and
so we should just ignore empty files.