net/core: Bug fix - buffer access synchronisation was incorrect.
Inside Endpoint the raw (write) callback incorrectly tried to send
any pending data from inside that call and then also continue to
allow new permit modifications to the input/output sequence of the
buffer via transfer_data(). This has been fixed and is also in fact
a slight improvement on performance - only transfer_data() is needed
as that takes care of sending any pending data as well as providing
access to new space in the output sequence of the buffer.