net/oop/Protocol: Add the protocol handler for OOP.
The protocol handler, which inherits of course from
the core ProtocolHandler class, understands how to
transform raw bytes from the Endpoint buffer into
a valid message.

Using the message factory it knows which message to
create based on an incoming ID in the header and
how to call the serialise method to ensure the
content of the message is deconstructed/reconstructed
for raw socket use.

It holds a queue of inbound and outbound messages.