net/core/ProtocolCreator: Add new base class of 'dialect'.
A protocol may, depending on component and context, have different dialects.
Of course, this certainly isn't true of simple client/server systems so we
make dialect an optional 2nd template parameter to Protocol (which still
inherits from ProtocolCreator).

If no dialect is needed the derived ProtocolHandler will still suffice as
before - it to however (the base class that is), now inherits from
ProtocolDialect.