Makefile: Add net/* objects as dependencies for sched/* objects.
The Makefile really needs cleaning up - well, the target
interdependencies. As far as the code goes not all test
apps depend on all project objects. The main problem is that
one unit in utility/ (NodeIO) has a code dependency on Node
in sched/ This must be fixed as utility is used by most tests
and components.