sched/Context: Provide a scheduling context to contain related objects.
A context holds the crawlers that connect, the server that talks 'OOP'
and of course the scheduler itself. It uses the Proactor to dispatch
the shcedulers run() method to a thread to handle it in the same way
the networking layer is executed. A breadcrumb trail is kept (though
not yet used) for each scheduler run.