sched/Context: Wait on all proactor threads (blocks) so as not to quit.
Because everything is non-blocking we have to prevent the Context and
therefore the Proactor objects from calling their destructors as this will
stop the io_service and join all threads and thus eventually terminating
the application. A call to start_crawling() should really block until the
program is sent a TERM signal.