sched/Scheduler: Bug fix - only defer a crawl when assigned > 0
Previously we had been adding crawlers to the deferred_crawls vector even
though the no. of assigned URLs returned from fill_crawler() was 0. This
ended up sending duplicate crawl data because currently the crawlers internal
copy of assigned URLs is not cleared after sending the data (using OOP) to the
remote crawler. This is something to consider later as we have to think about
fault tolerance in the cases where URLs assigned to a crawler may in fact not
get crawled because of a myriad of reasons.