net/oop/SchedulerCrawler: Add some flesh to the EndCrawl continue_dialog().
Send the data back to the scheduler after the BeginCrawl (so fake the actual
crawl) and handle it in the EndCrawl continue_dialog() method. It is here we
use DeferredUpdate or a Deferable rather, to update the schedule.

We defer the deletion of some messages until we're notified (the
GarbageCollector is a Subscriber) by the data processor that it's
finished.