net/core: Bug fix - must ensure that unset data in the buffer is sent.
When raw_send_callback() is dispatched there is a chance that the number of
bytes actually sent is less than the number we requested from the buffer.
Here, we fix this by allowing a pending no. of bytes to be given as a
parameter to transfer_data(). The pending no. of bytes is returned by
the consume_buffer() call already present in the ASIO callback handler for
send().