utility/Publisher: Actually, prefer our own Ref class over cref.
It is simple, we don't need a template and we can instantiate an
object of it at the beginning of the function rather than at
every iteration of the loop.

We can pass the same Ref to the blocking receive() call too due to the
cast operator provided by Ref.