utility/event: Bug fix - Events are destroyed before begin delivered.
When events are published asynchronously by a Proactor occasionally, well,
often, an event may be destroyed before all subscribers have processed the
event. We see this in the oodles component code. To overcome this problem
we must take a copy of the event object before publishing it - hence the
clone(). This is the only copy made. From here the object, wrapped in a boost
shared_ptr, is reference counted and destroyed when the last subscriber has
finished with it (falls out of the receive() scope).