utility/Event: Bug fix - Do not copy the list of subscribers.
Do not copy the subscribers. Only Publisher::broadcast() should make
a copy/clone of the Event base class. The source Event, 'e', remains
the authority as far as subscriptions are concerned. When it ('e')
goes out of scope or is deleted then the subcribers are removed according
to the destructor, ~Event().