sched/Node: Add correct normalise() method and propagate 'minimum'
Remove the templated normalise method from math (it was incorrect
and I cannot remember where the formula came from). Replace it
with a local implementation that does work. To ensure that negative
values don't create an imbalance propogate the minimum measure to
normalise() so that we may use it in the formula to ensure all
values are positive between 0 and 1.

The weight calculation formula has been simplified (yet again) as
it is still, I think, inaccurate and should returned to again when
we have real data (new links) coming in.