net/oop/Accessor: Add method for permitting serialisation of private data.
Objects that need to be transferred across the wire need to be serialised.
Normally attributes have private visibility and therefore a serialisation
method needs to be given special access to those attributes. Including
this file and using the PERMIT_SERIALISER_ACCESS macro achieves this.

We use our own accessor because boost::serialisation non-intrusive
'split' load/save methods cannot be granted access to private
members. So we wrap it up in a class and use that almost like a
proxy.