net/oop/Messages: Transfer/serialise the full-blown URL class.
The time has come to stop picking parts of the URL and sending those parts.
The crawler will now need to use most of the structure (parsed form of the
URL) and therefore we can justify sending the additional data.

To serialise the URL we must, however, enable the non-intrusive support for
it's serialise save/load methods. We do so by including the 'Boost.hpp' header
where it holds part of the solution (Accessor holds the rest).

We hint that the next patch may add support for serialising pointers to the
URL so as to rid the OOP layer of an additional transitory copy of the
n urls it will send.