net/core: Introduce OnConnect object.
I rather dislike this patch. It seems dirty and a bit
of a fudge/work-around. However, it is written - and
sometime ago in fact, before Christmas!

Anyway, the OnConnect object is used to begin work
once a successful connection has been established.
Connection establishment is burried deep within the
event-driven network layer so we must use this object
to defer the work and contain it in an object which
can be called via operator() when ready.

In crawl/Context we now use an OnConnect to begin
the dialog when a connection is established. Now,
the crawler and scheduler both use Context in a
similar way.