net/http: Initial commit of simple HTTP messaging interface
Untested this simple (not yet fully compliant) interface to the HTTP message
based protocol supports both the Response and Request, Client and Server
calls necessary to create HTTP chatter! Though not yet optimisied it is
(hopefully) efficient enough to assure transfer and parsing (deconstruct and
reconstruct) speeds. Only RFC2616 has been used as a reference.

The interface sports one fairly obvious difference - it has no buffer space
for the body of a message. Instead it accepts a file descriptor and body size
as parameters to the initial construction of a message hinting that
potentially on UNIX-like OSs sockets, pipes, files and shared memory may
be used to read/write the body. This idea is not yet tested ;)