net/core: Move to using Dispatcher internally to core networking
Replacing the direct use of io_service we use our own Dispatcher
composite of io_service and thread_pool.

Not sure yet whether this patch is necessary but it does simplify
a move to a different threading model later if we just pass around
a Dispatcher object to begin with.