net/core/SessionHandler: Introduce minimal SessionHandler object
This should replace both Dialect, which was actually session
dialog!, and OnConnect(), which will be affected by
SessionHandler::start().

ProtocolHandler and SessionHandler are the two Handlers
that the user (programmer) must define themselves and are
used together with the Endpoint to supply a meaningful network
communication flow. The least we can do is use a common suffix!