net/core/Client: Refactor Client to use SessionHandler.
First, we remove the Link system as we did with Endpoint. Second, we pair
a pending connection to the session passed in with the service to connect
to. The caller still retains ownership of the SessionHandler object passed
in (this is quite key to the refactor!) but it is paired with a
connection endpoint as it is passed through the async callbacks.