net/core/HandlerCreator: Take a caller context as the constructor parameter.
Here we maintain a reference to the caller context. The handler creator
is a constructor parameter to the Client or Server session handlers and
they maintain a reference to this handler creator. Therefore we can
keep the core net interface simple by ensuring the caller provides their
context when defining and declaring their handler(s) - makes sense right?