net/core: Modify caller-context connect event to follow start() pattern
Pass a reference of the caller context to Endpoint::start() which in turns
sends it on through the set Session. Session calls it's start() method and
passes itself on through back to the caller context. Sounds a bit convoluted
but it better fits the existing pattern of using start() on all components
involved and also Endpoint being the resource manager of Session and Protocol
etc.

Now, instead of connect_event, you must override start() in your supplied
caller context definition.