net/oop: Replace SchedulerCrawler with base class Session.
Session only defines a minimal interface - a derived Session must
override handle_message() as it is this method that is called on
continuously by handle_messages() until all are processed. Call
push_message() to push a message ready for sending.

All pointers should point to new-allocated (HEAP) messages and
ownership is transferred to the networking layer upon using
push_message() and, conversely, ownership is transferred to
the handle_message() implementor from the networking layer.