sched/Session: Add new OOP Session for the scheduler context.
This is more-or-less the old SchedulerCrawler dialect but with the crawler
context dialog stripped out. This is code necessary for the scheduler only
and the crawl namespace/context will hold it's own. This makes it much easier
to read and more light weight for each component.

It derives from the common OOP session which deals with the raw message
sending and receiving inc. the serialisation - this Session merely deals
with the specific dialog between the scheduler and (currently) the crawler.