net/core/Endpoint: Refactor to make set_* methods private and add friends.
First, always call socket() to retrieve the internal ASIO socket. Use
this as our single interface. Also, it can be made private due to the
refactor in this patch...

Though I don't want to make every core networking class a friend (and I'm
not!) I do want to be sure that the set_* methods are not public. I've made
them private and allowed only Client and Server as friends to be sure they can
call them as necessary - they are the only ones that should.