url/Parser: Replace boost spirit based parser in favour of easier bespoke one!
Revert to the older (wow - a year old!?) bespoke URL parser. It is simpler to
understand and bugfix, leaner, quicker to compile (I develop on a 12" G4
PowerBook mostly!) and probably faster too.

The main reason for this is that my spirit parser grammar was incorrect and
couldn't distinguish between a short hostname (no domain suffix) followed by
a port number or a username + password. Whats worse is I couldn't find out how
to fix it!

It's only a URL afterall - no need for something as full-featured as Spirit
here, I don't think!